# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger API for Golang Project Exchanger",
        "title": "Exchanger",
        "contact": {
            "name": "Ilya",
            "email": "biv_1998@mail.ru"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/currencies/": {
            "get": {
                "description": "Get name, description and abbreviation for all currencies in database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Currencies"
                ],
                "summary": "currencies",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Currency"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorStruct"
                        }
                    }
                }
            }
        },
        "/rates/": {
            "get": {
                "description": "Get name, description, abbreviation for currency, name, description and abbreviation for base currency, time and rate",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rates"
                ],
                "summary": "rates",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.ExchangeRate"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorStruct"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorStruct"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.Currency": {
            "type": "object",
            "properties": {
                "abbreviation": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "api.ErrorStruct": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "api.ExchangeRate": {
            "type": "object",
            "properties": {
                "base_currency": {
                    "$ref": "#/definitions/api.Currency"
                },
                "currency": {
                    "$ref": "#/definitions/api.Currency"
                },
                "id": {
                    "type": "string"
                },
                "rate": {
                    "type": "number"
                },
                "time": {
                    "type": "string"
                }
            }
        }
    }
}